backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.5

# Google Cloud dependencies
google-cloud-firestore==2.18.0
google-cloud-storage==2.17.0
google-cloud-pubsub==2.23.0
firebase-admin==6.5.0

# Additional dependencies for AI and async processing
openai==1.54.4
aiohttp==3.10.10
python-jose[cryptography]==3.3.0
httpx==0.27.2
